ICU-13197 catch one more edge case where a mapping has no compose boundary after it
X-SVN-Rev: 40156

+++ icu4c/source/tools/gennorm2/norms.cpp
@@ -131,7 +131,7 @@ void Norms::reorder(UnicodeString &mapping, BuilderReorderingBuffer &buffer) con
     }
 }
 
-UBool Norms::combinesWithCCBetween(const Norm &norm, uint8_t lowCC, uint8_t highCC) const {
+UBool Norms::combinesWithCCBetween(const Norm &norm, uint8_t lowCC, int32_t highCC) const {
     if((highCC-lowCC)>=2) {
         int32_t length;
         const CompositionPair *pairs=norm.getCompositionPairs(length);
